refactor: v2 이전의 API를 제공하는 컨트롤러들 제거
